refactor(shared): move test utils to shared module
closes #8
