+++ src/modules/scheduler/application/calendar/lessons.loader.test.ts
@@ -18,7 +18,7 @@ import {
 import { DateTime } from 'luxon';
 import { LessonsLoader } from './lessons.loader';
 import { MikroORM } from '@mikro-orm/postgresql';
-import { testMikroormProvider } from '../../../../../test-utils';
+import { testMikroormProvider } from '../../../shared/tests';
 
 describe('LessonsLoader', () => {
   let loader: LessonsLoader;
