+++ src/modules/scheduler/application/calendar/subject-versions.loader.test.ts
@@ -15,7 +15,7 @@ import {
 import { DateTime } from 'luxon';
 import { SubjectVersionsLoader } from './subject-versions.loader';
 import { MikroORM } from '@mikro-orm/postgresql';
-import { testMikroormProvider } from '../../../../../test-utils';
+import { testMikroormProvider } from '../../../shared/tests';
 
 describe('SubjectVersionsLoader', () => {
   let loader: SubjectVersionsLoader;
